+++ src/max-plugin/package.json
@@ -3,7 +3,8 @@
   "version": "0.0.0",
   "private": true,
   "scripts": {
-    "start": "node max_sockets.js"
+    "start": "node max_sockets.js",
+    "serve": "node server.js"
   },
   "dependencies": {
     "body-parser": "~1.17.1",
@@ -11,6 +12,7 @@
     "debug": "~2.6.3",
     "ejs": "~2.5.6",
     "express": "~4.15.2",
+    "max-api": "^1.0.1",
     "morgan": "~1.8.1",
     "serve-favicon": "~2.4.2",
     "ws": "^4.0.0"
